Refactor IntakeCommand as a command group, add commands to control intake tasks.
